# syncify.web.client/src/api/generated/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Syncify.Web.Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/calendar-events": {
      "post": {
        "tags": [
          "CalendarEvent"
        ],
        "operationId": "CreateCalendarEvent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarEventCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarEventCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarEventCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarEventGetDto>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarEventGetDto>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarEventGetDto>"
                }
              }
            }
          }
        }
      }
    },
    "/api/calendar-events/calendar/{calendarId}": {
      "get": {
        "tags": [
          "CalendarEvent"
        ],
        "operationId": "GetCalendarEvents",
        "parameters": [
          {
            "name": "calendarId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<List<CalendarEventGetDto>>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<List<CalendarEventGetDto>>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<List<CalendarEventGetDto>>"
                }
              }
            }
          }
        }
      }
    },
    "/api/calendar-events/{id}": {
      "get": {
        "tags": [
          "CalendarEvent"
        ],
        "operationId": "GetCalendarEventById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarEventGetDto>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarEventGetDto>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarEventGetDto>"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CalendarEvent"
        ],
        "operationId": "UpdateCalendarEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarEventUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarEventUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarEventUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarEventGetDto>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarEventGetDto>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarEventGetDto>"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CalendarEvent"
        ],
        "operationId": "DeleteCalendarEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/calendars/{id}": {
      "get": {
        "tags": [
          "Calendars"
        ],
        "operationId": "GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarGetDto>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarGetDto>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarGetDto>"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Calendars"
        ],
        "operationId": "Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarGetDto>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarGetDto>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<CalendarGetDto>"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Calendars"
        ],
        "operationId": "Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/recipes/{id}": {
      "get": {
        "tags": [
          "Recipes"
        ],
        "operationId": "GetRecipeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeGetDto>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeGetDto>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeGetDto>"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Recipes"
        ],
        "operationId": "DeleteRecipe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<Boolean>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<Boolean>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<Boolean>"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipetags": {
      "get": {
        "tags": [
          "RecipeTags"
        ],
        "operationId": "GetTags",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<List<RecipeTagDto>>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<List<RecipeTagDto>>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<List<RecipeTagDto>>"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RecipeTags"
        ],
        "operationId": "CreateTag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeTagCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeTagCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeTagCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeTagDto>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeTagDto>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeTagDto>"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipetags/{id}": {
      "get": {
        "tags": [
          "RecipeTags"
        ],
        "operationId": "GetTagById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeTagDto>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeTagDto>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeTagDto>"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RecipeTags"
        ],
        "operationId": "UpdateTag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeTagCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeTagCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeTagCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeTagDto>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeTagDto>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeTagDto>"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "RecipeTags"
        ],
        "operationId": "DeleteTag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<Boolean>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<Boolean>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<Boolean>"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/recipe-ingredient/recipe/{recipeId}": {
      "get": {
        "tags": [
          "RecipieIngredient"
        ],
        "operationId": "GetAll",
        "parameters": [
          {
            "name": "recipeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<List<RecipeIngredientGetDto>>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<List<RecipeIngredientGetDto>>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<List<RecipeIngredientGetDto>>"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalendarEventCreateDto": {
        "required": [
          "calendarEventType",
          "calendarId",
          "recurrenceType",
          "startDate",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "displayColor": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string",
            "format": "time",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "time",
            "nullable": true
          },
          "calendarEventType": {
            "$ref": "#/components/schemas/CalendarEventType"
          },
          "recurrenceType": {
            "$ref": "#/components/schemas/RecurrenceType"
          },
          "calendarId": {
            "type": "integer",
            "format": "int32"
          },
          "recurrenceWeekDays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayOfWeek"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CalendarEventGetDto": {
        "required": [
          "calendarEventType",
          "calendarId",
          "id",
          "recurrenceType",
          "startDate",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "displayColor": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string",
            "format": "time",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "time",
            "nullable": true
          },
          "calendarEventType": {
            "$ref": "#/components/schemas/CalendarEventType"
          },
          "recurrenceType": {
            "$ref": "#/components/schemas/RecurrenceType"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "calendarId": {
            "type": "integer",
            "format": "int32"
          },
          "recurrenceWeekDays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayOfWeek"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CalendarEventType": {
        "enum": [
          "Event",
          "Task"
        ],
        "type": "string",
        "format": "int32"
      },
      "CalendarEventUpdateDto": {
        "required": [
          "calendarEventType",
          "recurrenceType",
          "startDate",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "displayColor": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string",
            "format": "time",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "time",
            "nullable": true
          },
          "calendarEventType": {
            "$ref": "#/components/schemas/CalendarEventType"
          },
          "recurrenceType": {
            "$ref": "#/components/schemas/RecurrenceType"
          },
          "recurrenceWeekDays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayOfWeek"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CalendarGetDto": {
        "required": [
          "createdByUserId",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "createdByUserId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CalendarUpdateDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "DayOfWeek": {
        "enum": [
          "Sunday",
          "Monday",
          "Tuesday",
          "Wednesday",
          "Thursday",
          "Friday",
          "Satuday"
        ],
        "type": "string"
      },
      "EmptyResponse": {
        "required": [
          "errors",
          "hasErrors"
        ],
        "type": "object",
        "properties": {
          "hasErrors": {
            "type": "boolean",
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "additionalProperties": false
      },
      "Error": {
        "required": [
          "errorMessage"
        ],
        "type": "object",
        "properties": {
          "errorMessage": {
            "type": "string"
          },
          "propertyName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RecipeTagCreateDto": {
        "required": [
          "name",
          "recipeId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "recipeId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "RecipeTagDto": {
        "required": [
          "id",
          "name",
          "recipeId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "recipeId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "RecurrenceType": {
        "enum": [
          "Daily",
          "Weekly",
          "BiWeekly",
          "Monthly",
          "Custom"
        ],
        "type": "string"
      },
      "Response<Boolean>": {
        "required": [
          "data",
          "errors",
          "hasErrors"
        ],
        "type": "object",
        "properties": {
          "hasErrors": {
            "type": "boolean",
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "data": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "Response<CalendarEventGetDto>": {
        "required": [
          "data",
          "errors",
          "hasErrors"
        ],
        "type": "object",
        "properties": {
          "hasErrors": {
            "type": "boolean",
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "data": {
            "$ref": "#/components/schemas/CalendarEventGetDto"
          }
        },
        "additionalProperties": false
      },
      "Response<CalendarGetDto>": {
        "required": [
          "data",
          "errors",
          "hasErrors"
        ],
        "type": "object",
        "properties": {
          "hasErrors": {
            "type": "boolean",
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "data": {
            "$ref": "#/components/schemas/CalendarGetDto"
          }
        },
        "additionalProperties": false
      },
      "Response<List<CalendarEventGetDto>>": {
        "required": [
          "errors",
          "hasErrors"
        ],
        "type": "object",
        "properties": {
          "hasErrors": {
            "type": "boolean",
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalendarEventGetDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Response<List<RecipeIngredientGetDto>>": {
        "required": [
          "errors",
          "hasErrors"
        ],
        "type": "object",
        "properties": {
          "hasErrors": {
            "type": "boolean",
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeIngredientGetDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Response<List<RecipeTagDto>>": {
        "required": [
          "errors",
          "hasErrors"
        ],
        "type": "object",
        "properties": {
          "hasErrors": {
            "type": "boolean",
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeTagDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Response<RecipeGetDto>": {
        "required": [
          "data",
          "errors",
          "hasErrors"
        ],
        "type": "object",
        "properties": {
          "hasErrors": {
            "type": "boolean",
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "data": {
            "$ref": "#/components/schemas/RecipeGetDto"
          }
        },
        "additionalProperties": false
      },
...
      "Response<RecipeTagDto>": {
        "required": [
          "data",
          "errors",
          "hasErrors"
        ],
        "type": "object",
        "properties": {
          "hasErrors": {
            "type": "boolean",
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "data": {
            "$ref": "#/components/schemas/RecipeTagDto"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}